deleted unnecessary files and code. Refactored incorrectly named files and strings
