+++ app/src/main/AndroidManifest.xml
@@ -31,13 +31,9 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".ActivityRankings"
+            android:name=".ActivityMain"
             android:label="Activity Rankings"
             android:windowSoftInputMode="adjustResize" />
-        <activity
-            android:name=".ActivitySelection"
-            android:label="Activity Selection"
-            android:parentActivityName=".ActivityRankings" />
         <activity android:name=".Login">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -51,40 +47,15 @@
         <activity
             android:name=".ColorPicker"
             android:label="Color Picker"
-            android:parentActivityName=".ActivityRankings" />
+            android:parentActivityName=".ActivityMain" />
         <activity
-            android:name=".SportSelection"
-            android:label="@string/title_activity_sport_selection"
+            android:name=".ActivitySelection"
+            android:label="@string/label_activity_selection"
             android:parentActivityName=".Login"
             android:theme="@style/Theme.AppCompat.DayNight.NoActionBar" />
         <activity
-            android:name=".TestGETBackEnd"
-            android:label="Test GET Back End"
-            android:parentActivityName=".ActivityRankings" />
-        <activity
-            android:name=".TestPOSTPUTSportBackEnd"
-            android:label="Test POST PUT Sport Back End"
-            android:parentActivityName=".ActivityRankings" />
-        <activity
-            android:name=".TestPOSTPUTPlayerBackEnd"
-            android:label="Test POST PUT Player Back End"
-            android:parentActivityName=".ActivityRankings" />
-        <activity
-            android:name=".TestPOSTPUTFollowBackEnd"
-            android:label="Test POST PUT Follow Back End"
-            android:parentActivityName=".ActivityRankings" />
-        <activity
-            android:name=".TestPOSTPUTMatchBackEnd"
-            android:label="Test POST PUT Match Back End"
-            android:parentActivityName=".ActivityRankings"></activity>
-        <activity
-            android:name=".TestDELETEBackEnd"
-            android:label="Test DELETE Back End"
-            android:parentActivityName=".ActivityRankings"></activity>
-        <activity
             android:name=".OnlineHelpSystem"
             android:label="Online Help System"
-            android:parentActivityName=".ActivityRankings"></activity>
+            android:parentActivityName=".ActivityMain"></activity>
     </application>
-
 </manifest>
