--- app/src/main/java/edu/calvin/cs262/pilot/knightrank/ChallengeConfirmation.java
@@ -1,115 +1,0 @@
-package edu.calvin.cs262.pilot.knightrank;
-
-import android.content.Context;
-import android.content.SharedPreferences;
-import android.graphics.Color;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.os.Bundle;
-import android.util.Log;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.ArrayAdapter;
-import android.widget.ListView;
-import android.widget.RelativeLayout;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Objects;
-
-public class ChallengeConfirmation extends Fragment {
-
-    //Class variables.
-    private static final String LOG_TAG =
-            ChallengeConfirmation.class.getSimpleName();
-
-    // For use with shared preferences.
-    private static String PLACEHOLDER2 = "";
-
-    // Share preferences file (custom)
-    private SharedPreferences mPreferences;
-    // Shared preferences file (default)
-    private SharedPreferences mPreferencesDefault;
-
-    // Name of the custom shared preferences file.
-    private static final String sharedPrefFile = "pilot.cs262.calvin.edu.knightrank";
-
-    private ListView mChallengeResults;
-
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-
-        super.onCreate(savedInstanceState);
-
-        // Set shared preferences component.
-        // Note: modified from the one in activities as this is a fragment.
-        mPreferences = Objects.requireNonNull(this.getActivity()).getSharedPreferences(sharedPrefFile, Context.MODE_PRIVATE);
-        mPreferencesDefault = android.support.v7.preference.PreferenceManager.getDefaultSharedPreferences(this.getActivity());
-
-        // Placeholder code as example of how to get values from the default SharedPrefs file.
-        String syncFreq = mPreferencesDefault.getString(SettingsActivity.KEY_SYNC_FREQUENCY, "-1");
-
-        // Placeholder code as example of how to restore values to UI components from shared preferences.
-        //username_main.setText(mPreferences.getString(USER_NAME, ""));
-        //password_main.setText(mPreferences.getString(USER_PASSWORD, ""));
-    }
-
-    @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-
-
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_challenge_confirmation, container, false);
-    }
-
-    /**
-     * Method implemented to change background color programmatically via color picket.
-     *
-     * @param view object
-     * @param savedInstanceState object
-     */
-    @Override
-    public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
-        super.onViewCreated(view, savedInstanceState);
-
-        // Change the background color to what was selected in color picker.
-        // Note: Change color by using findViewById and ID of the UI element you wish to change.
-        RelativeLayout thisLayout = Objects.requireNonNull(getView()).findViewById(R.id.fragment_challenge_confirmation_root_layout);
-        thisLayout.setBackgroundColor(mPreferences.getInt(ColorPicker.APP_BACKGROUND_COLOR_ARGB, Color.WHITE));
-
-        mChallengeResults = (ListView) getView().findViewById(R.id.challenge_confirmation_listview);
-        List<String> challenge_results_arraylist = new ArrayList<String>();
-        /*
-        Theoretical entries, as we don't have a backend #FIXME
-        */
-        challenge_results_arraylist.add("vs. Tommy, Smash 4");
-        challenge_results_arraylist.add("vs. Samantha, Tekken 7");
-        ArrayAdapter<String> arrayAdapter = new ArrayAdapter<String>(
-                getActivity(),
-                android.R.layout.simple_list_item_1,
-                challenge_results_arraylist);
-        mChallengeResults.setAdapter(arrayAdapter);
-
-        int value = mPreferences.getInt(ColorPicker.APP_BACKGROUND_COLOR_ARGB, Color.BLACK);
-
-        Log.e(LOG_TAG,"Value of color is: " + value);
-    }
-
-    /**
-     * Method currently called to store values to shared preferences file.
-     */
-    @Override
-    public void onPause() {
-        super.onPause();
-
-        SharedPreferences.Editor preferencesEditor4 = mPreferences.edit();
-
-        preferencesEditor4.putString(PLACEHOLDER2, "Placeholder text 2");
-
-        preferencesEditor4.apply();
-    }
-}
