+++ app/src/main/java/edu/calvin/cs262/pilot/knightrank/ConfirmMatch.java
@@ -14,22 +14,19 @@ import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
 import android.widget.ListView;
 import android.widget.RelativeLayout;
-import android.widget.TextView;
-
-import org.w3c.dom.Text;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
-public class MyStats extends Fragment {
+public class ConfirmMatch extends Fragment {
 
     //Class variables.
     private static final String LOG_TAG =
-            MyStats.class.getSimpleName();
+            ConfirmMatch.class.getSimpleName();
 
     // For use with shared preferences.
-    private static String PLACEHOLDER6 = "";
+    private static String PLACEHOLDER2 = "";
 
     // Share preferences file (custom)
     private SharedPreferences mPreferences;
@@ -39,9 +36,8 @@ public class MyStats extends Fragment {
     // Name of the custom shared preferences file.
     private static final String sharedPrefFile = "pilot.cs262.calvin.edu.knightrank";
 
-    private TextView winStat;
-    private TextView lossStat;
-    private ListView mUpcomingMatchesList;
+    private ListView mChallengeResults;
+
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
@@ -65,8 +61,9 @@ public class MyStats extends Fragment {
     public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
 
+
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_my_stats, container, false);
+        return inflater.inflate(R.layout.fragment_confirm_match, container, false);
     }
 
     /**
@@ -81,29 +78,23 @@ public class MyStats extends Fragment {
 
         // Change the background color to what was selected in color picker.
         // Note: Change color by using findViewById and ID of the UI element you wish to change.
-        RelativeLayout thisLayout = Objects.requireNonNull(getView()).findViewById(R.id.fragment_my_rankings_root_layout);
+        RelativeLayout thisLayout = Objects.requireNonNull(getView()).findViewById(R.id.fragment_challenge_confirmation_root_layout);
         thisLayout.setBackgroundColor(mPreferences.getInt(ColorPicker.APP_BACKGROUND_COLOR_ARGB, Color.WHITE));
 
-        int value = mPreferences.getInt(ColorPicker.APP_BACKGROUND_COLOR_ARGB, Color.BLACK);
-
-        lossStat = (TextView) getView().findViewById(R.id.loss_stat_textview);
-        winStat = (TextView) getView().findViewById(R.id.win_stat_textview);
-        mUpcomingMatchesList = (ListView) getView().findViewById(R.id.my_stats_upcoming_matches_listview);
-
-        List<String> upcomingMatchesArrayList = new ArrayList<String>();
-
-        upcomingMatchesArrayList.add("vs. Tom, Injustice 2");
-        upcomingMatchesArrayList.add("vs. Jerry, Mortal Kombat X");
-        upcomingMatchesArrayList.add("vs. Bill, Super Smash Bros Melee");
-        upcomingMatchesArrayList.add("vs. Ted, Soulcaliber V");
-
+        mChallengeResults = (ListView) getView().findViewById(R.id.challenge_confirmation_listview);
+        List<String> challenge_results_arraylist = new ArrayList<String>();
+        /*
+        Theoretical entries, as we don't have a backend #FIXME
+        */
+        challenge_results_arraylist.add("vs. Tommy, Smash 4");
+        challenge_results_arraylist.add("vs. Samantha, Tekken 7");
         ArrayAdapter<String> arrayAdapter = new ArrayAdapter<String>(
                 getActivity(),
                 android.R.layout.simple_list_item_1,
-                upcomingMatchesArrayList);
+                challenge_results_arraylist);
+        mChallengeResults.setAdapter(arrayAdapter);
 
-
-        mUpcomingMatchesList.setAdapter(arrayAdapter);
+        int value = mPreferences.getInt(ColorPicker.APP_BACKGROUND_COLOR_ARGB, Color.BLACK);
 
         Log.e(LOG_TAG,"Value of color is: " + value);
     }
@@ -115,10 +106,10 @@ public class MyStats extends Fragment {
     public void onPause() {
         super.onPause();
 
-        SharedPreferences.Editor preferencesEditor8 = mPreferences.edit();
+        SharedPreferences.Editor preferencesEditor4 = mPreferences.edit();
 
-        preferencesEditor8.putString(PLACEHOLDER6, "Placeholder text 6");
+        preferencesEditor4.putString(PLACEHOLDER2, "Placeholder text 2");
 
-        preferencesEditor8.apply();
+        preferencesEditor4.apply();
     }
 }
