+++ app/src/main/java/edu/calvin/cs262/pilot/knightrank/PastChallenges.java
@@ -105,7 +105,7 @@ public class PastChallenges extends Fragment {
     public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_past_challenges, container, false);
+        return inflater.inflate(R.layout.fragment_match_history, container, false);
     }
 
     /**
