+++ app/src/main/res/layout/activity_color_picker.xml
@@ -16,28 +16,6 @@
         android:theme="@style/ThemeOverlay.AppCompat.ActionBar"
         app:popupTheme="@style/ThemeOverlay.AppCompat.Light"/>
 
-    <!--
-    <TextView
-        android:id="@+id/labelA"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/my_toolbar"
-        android:layout_alignParentStart="true"
-        android:text="A"
-        android:textAppearance="?android:attr/textAppearanceMedium" />
-
-    <SeekBar
-        android:id="@+id/seekA"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_alignBottom="@id/labelA"
-        android:layout_marginLeft="-3dp"
-        android:layout_marginBottom="0dp"
-        android:layout_toRightOf="@id/labelA"
-        android:max="255"
-        android:progress="255" />
-        -->
-
     <TextView
         android:id="@+id/labelR"
         android:layout_width="wrap_content"
@@ -46,12 +24,6 @@
         android:layout_alignParentStart="true"
         android:text="R"
         android:textAppearance="?android:attr/textAppearanceMedium" />
-
-    <!--
-    android:layout_below="@id/labelA"
-    android:layout_alignParentStart="true"
-    android:layout_marginTop="0dp"
-    -->
 
     <SeekBar
         android:id="@+id/seekR"
@@ -64,10 +36,6 @@
         android:layout_toRightOf="@+id/labelR"
         android:max="255"
         android:progress="255" />
-
-    <!--
-    android:layout_below="@+id/seekA"
-    -->
 
     <TextView
         android:id="@+id/labelG"
@@ -161,6 +129,5 @@
                     android:text="App Background Color" />
             </LinearLayout>
         </ScrollView>
-
     </LinearLayout>
 </RelativeLayout>
