+++ app/src/main/res/layout/activity_online_help_system.xml
@@ -1,5 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
-<!-- Use DrawerLayout as root container for activity -->
 <android.support.v4.widget.DrawerLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
@@ -38,10 +37,8 @@
             android:layout_height="wrap_content"
             android:text="Choose help topic using drawer navigation" />
     </FrameLayout>
-
 </LinearLayout>
 
-    <!-- Container for contents of drawer - use NavigationView to make configuration easier -->
     <android.support.design.widget.NavigationView
         android:id="@+id/help_nav_view"
         android:layout_width="wrap_content"
@@ -50,5 +47,4 @@
         android:fitsSystemWindows="true"
         app:menu="@menu/help_drawer_view"
         app:headerLayout="@layout/nav_header" />
-
 </android.support.v4.widget.DrawerLayout>
