+++ app/src/main/res/layout/activity_selection.xml
@@ -31,122 +31,11 @@
         android:theme="@style/ThemeOverlay.AppCompat.ActionBar"
         app:popupTheme="@style/ThemeOverlay.AppCompat.Light"/>
 
-    <!--
-    <TextView
-        android:id="@+id/activity_selection_textview"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="55dp"
-        android:layout_marginLeft="10dp"
-        android:text="@string/title_activity_selection"
-        android:textColor="#800000"
-        android:textSize="44sp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintHorizontal_bias="0.529"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.139" />
-        -->
-
     <ListView
         android:id="@+id/activity_selection_listview"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginTop="55dp"
         android:layout_above="@+id/return_to_activity_rankings_button">
-
     </ListView>
-    <!--
-    android:layout_below="@+id/activity_selection_textview"
-    -->
-
-    <!--
-    <ScrollView
-        android:layout_width="368dp"
-        android:layout_height="360dp"
-        android:layout_marginTop="120dp"
-        android:layout_marginLeft="8dp"
-        tools:layout_editor_absoluteX="8dp"
-        tools:layout_editor_absoluteY="143dp">
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="vertical" >
-
-            <Button
-                android:id="@+id/button"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:onClick="launchActivity"
-                android:text="Chess" />
-
-            <Button
-                android:id="@+id/button2"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:onClick="launchActivity"
-                android:text="Football" />
-
-            <Button
-                android:id="@+id/button3"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:onClick="launchActivity"
-                android:text="World of Warcraft Arena" />
-
-            <Button
-                android:id="@+id/button4"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:onClick="launchActivity"
-                android:text="DOTA 2" />
-
-            <Button
-                android:id="@+id/button5"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:onClick="launchActivity"
-                android:text="PUBG" />
-
-            <Button
-                android:id="@+id/button6"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:onClick="launchActivity"
-                android:text="Volleyball" />
-
-            <Button
-                android:id="@+id/button7"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:onClick="launchActivity"
-                android:text="Overwatch" />
-
-            <Button
-                android:id="@+id/button8"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:onClick="launchActivity"
-                android:text="Starcraft 2" />
-
-            <Button
-                android:id="@+id/button9"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:onClick="launchActivity"
-                android:text="Tom Clancy's Rainbow Six Siege" />
-
-            <Button
-                android:id="@+id/button10"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:onClick="launchActivity"
-                android:text="Randomized" />
-        </LinearLayout>
-    </ScrollView>
-    
-   -->
-
 </RelativeLayout>
