+++ app/src/main/res/layout/activity_sport_selection.xml
@@ -5,9 +5,9 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/activity_sport_selection_root_layout"
-    tools:context="edu.calvin.cs262.pilot.knightrank.SportSelection">
+    tools:context="edu.calvin.cs262.pilot.knightrank.ActivitySelection">
 
-    <include layout="@layout/content_sport_selection" />
+    <include layout="@layout/content_activity_selection" />
 
     <android.support.design.widget.FloatingActionButton
         android:id="@+id/fab"
@@ -25,5 +25,4 @@
         android:elevation="4dp"
         android:theme="@style/ThemeOverlay.AppCompat.ActionBar"
         app:popupTheme="@style/ThemeOverlay.AppCompat.Light"/>
-
 </android.support.design.widget.CoordinatorLayout>
