--- app/src/main/res/layout/activity_test_postfollow_back_end.xml
@@ -1,186 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:id="@+id/activity_test_follow_post_back_end_root_layout"
-    tools:context="edu.calvin.cs262.pilot.knightrank.TestPOSTPUTFollowBackEnd">
-
-    <Button
-        android:id="@+id/send_follow_put"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignEnd="@+id/request_status"
-        android:layout_alignBottom="@+id/network_connected"
-        android:layout_marginEnd="74dp"
-        android:layout_marginBottom="-17dp"
-        android:onClick="put"
-        android:text="PUT" />
-
-    <EditText
-        android:id="@+id/input_follow_put_id"
-        android:layout_width="361dp"
-        android:layout_height="52dp"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginStart="0dp"
-        android:layout_marginTop="136dp"
-        android:autofillHints=""
-        android:hint="Enter a id (integer):"
-        android:inputType="text"
-        android:textSize="12sp" />
-
-    <TextView
-        android:id="@+id/follow_put_id"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginStart="0dp"
-        android:layout_marginTop="114dp"
-        android:text="Follow ID:" />
-
-    <TextView
-        android:id="@+id/post_follow_note_2"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentTop="true"
-        android:layout_marginStart="8dp"
-        android:layout_marginTop="409dp"
-        android:layout_marginEnd="8dp"
-        android:text="Bug: Click into any field first time after loading into activity before hitting post, otherwise activity crashes."
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/search_string" />
-
-    <TextView
-        android:id="@+id/follow_note"
-        android:layout_width="332dp"
-        android:layout_height="wrap_content"
-        android:layout_alignParentTop="true"
-        android:layout_marginStart="-7dp"
-        android:layout_marginTop="87dp"
-        android:layout_toEndOf="@+id/follow_rank"
-        android:text="Note: Disable foreign constraint enforcement to test" />
-
-    <TextView
-        android:id="@+id/follow_rank"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginStart="0dp"
-        android:layout_marginTop="338dp"
-        android:text="Rank:" />
-
-    <EditText
-        android:id="@+id/input_follow_rank"
-        android:layout_width="361dp"
-        android:layout_height="52dp"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginStart="0dp"
-        android:layout_marginTop="359dp"
-        android:autofillHints=""
-        android:hint="Enter a rank (integer):"
-        android:inputType="text"
-        android:textSize="12sp" />
-
-    <android.support.v7.widget.Toolbar
-        android:id="@+id/my_toolbar"
-        android:layout_width="match_parent"
-        android:layout_height="?attr/actionBarSize"
-        android:background="?attr/colorPrimary"
-        android:elevation="4dp"
-        android:theme="@style/ThemeOverlay.AppCompat.ActionBar"
-        app:popupTheme="@style/ThemeOverlay.AppCompat.Light"/>
-
-    <TextView
-        android:id="@+id/post_put_follow"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentTop="true"
-        android:layout_centerHorizontal="true"
-        android:layout_marginTop="61dp"
-        android:text="POST or PUT a Follow" />
-
-    <TextView
-        android:id="@+id/follow_sport_id"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginStart="0dp"
-        android:layout_marginTop="188dp"
-        android:text="Sport ID:" />
-
-    <TextView
-        android:id="@+id/follow_player_id"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginStart="0dp"
-        android:layout_marginTop="262dp"
-        android:text="Player ID:" />
-
-
-    <EditText
-        android:id="@+id/input_follow_sport_id"
-        android:layout_width="361dp"
-        android:layout_height="52dp"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginStart="0dp"
-        android:layout_marginTop="208dp"
-        android:autofillHints=""
-        android:hint="Enter a sport ID (integer):"
-        android:inputType="text"
-        android:textSize="12sp" />
-
-    <EditText
-        android:id="@+id/input_follow_player_id"
-        android:layout_width="361dp"
-        android:layout_height="52dp"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginStart="0dp"
-        android:layout_marginTop="284dp"
-        android:autofillHints=""
-        android:hint="Enter a player ID (integer):"
-        android:inputType="text"
-        android:textSize="12sp" />
-
-    <Button
-        android:id="@+id/send_follow_post"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignEnd="@+id/request_status"
-        android:layout_alignBottom="@+id/network_connected"
-        android:layout_marginEnd="-30dp"
-        android:layout_marginBottom="-17dp"
-        android:onClick="post"
-        android:text="POST" />
-
-    <TextView
-        android:id="@+id/network_connected"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentBottom="true"
-        android:layout_marginStart="16dp"
-        android:layout_marginBottom="18dp"
-        android:text="Network Connectivity" />
-
-    <TextView
-        android:id="@+id/request_status"
-        android:layout_width="344dp"
-        android:layout_height="57dp"
-        android:layout_alignStart="@+id/network_connected"
-        android:layout_alignParentBottom="true"
-        android:layout_marginStart="-7dp"
-        android:layout_marginBottom="65dp"
-        android:text="Request Status:" />
-
-</RelativeLayout>
