--- app/src/main/res/layout/activity_test_postplayer_back_end.xml
@@ -1,167 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:id="@+id/activity_test_player_post_back_end_root_layout"
-    tools:context="edu.calvin.cs262.pilot.knightrank.TestPOSTPUTPlayerBackEnd">
-
-    <Button
-        android:id="@+id/send_player_put"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentEnd="true"
-        android:layout_alignParentBottom="true"
-        android:layout_marginEnd="96dp"
-        android:layout_marginBottom="0dp"
-        android:onClick="put"
-        android:text="PUT" />
-
-    <TextView
-        android:id="@+id/player_put_id"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginStart="0dp"
-        android:layout_marginTop="95dp"
-        android:text="Player ID:" />
-
-    <EditText
-        android:id="@+id/input_player_put_id"
-        android:layout_width="361dp"
-        android:layout_height="52dp"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginStart="0dp"
-        android:layout_marginTop="117dp"
-        android:autofillHints=""
-        android:hint="Enter a id (integer):"
-        android:inputType="text"
-        android:textSize="12sp" />
-
-    <TextView
-        android:id="@+id/post_player_note_2"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_alignParentEnd="true"
-        android:layout_marginStart="9dp"
-        android:layout_marginTop="365dp"
-        android:layout_marginEnd="7dp"
-        android:text="Bug: Click into any field first time after loading into activity before hitting post, otherwise activity crashes."
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/search_string" />
-
-    <android.support.v7.widget.Toolbar
-        android:id="@+id/my_toolbar"
-        android:layout_width="match_parent"
-        android:layout_height="?attr/actionBarSize"
-        android:background="?attr/colorPrimary"
-        android:elevation="4dp"
-        android:theme="@style/ThemeOverlay.AppCompat.ActionBar"
-        app:popupTheme="@style/ThemeOverlay.AppCompat.Light"/>
-
-    <TextView
-        android:id="@+id/post_put_player"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentTop="true"
-        android:layout_centerHorizontal="true"
-        android:layout_marginTop="61dp"
-        android:text="POST or PUT a Player" />
-
-    <TextView
-        android:id="@+id/player_email"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginStart="0dp"
-        android:layout_marginTop="173dp"
-        android:text="Player Email" />
-
-    <TextView
-        android:id="@+id/account_creation_date"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginStart="0dp"
-        android:layout_marginTop="251dp"
-        android:text="Account Creation Date (must match timestamp type)" />
-
-
-    <EditText
-        android:id="@+id/input_player_email"
-        android:layout_width="361dp"
-        android:layout_height="52dp"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginStart="0dp"
-        android:layout_marginTop="195dp"
-        android:autofillHints=""
-        android:hint="Enter a email (string):"
-        android:inputType="text"
-        android:textSize="12sp" />
-
-    <EditText
-        android:id="@+id/input_player_account_creation_date"
-        android:layout_width="361dp"
-        android:layout_height="52dp"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginStart="0dp"
-        android:layout_marginTop="305dp"
-        android:autofillHints=""
-        android:hint="Enter a account creation date (timestamp format)"
-        android:inputType="text"
-        android:textSize="12sp" />
-
-    <Button
-        android:id="@+id/send_player_post"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentEnd="true"
-        android:layout_alignParentBottom="true"
-        android:layout_marginEnd="0dp"
-        android:layout_marginBottom="0dp"
-        android:onClick="post"
-        android:text="POST" />
-
-    <TextView
-        android:id="@+id/network_connected"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentBottom="true"
-        android:layout_marginStart="16dp"
-        android:layout_marginBottom="18dp"
-        android:text="Network Connectivity" />
-
-    <TextView
-        android:id="@+id/request_status"
-        android:layout_width="344dp"
-        android:layout_height="57dp"
-        android:layout_alignStart="@+id/network_connected"
-        android:layout_alignParentEnd="true"
-        android:layout_alignParentBottom="true"
-        android:layout_marginStart="-1dp"
-        android:layout_marginEnd="24dp"
-        android:layout_marginBottom="102dp"
-        android:text="Request Status:" />
-
-    <TextView
-        android:id="@+id/player_timestamp_note"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginStart="0dp"
-        android:layout_marginTop="277dp"
-        android:text="Format: 2018-11-03 00:53:57.048546" />
-
-</RelativeLayout>
