--- app/src/main/res/layout/activity_test_postsport_back_end.xml
@@ -1,156 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:id="@+id/activity_test_sport_post_back_end_root_layout"
-    tools:context="edu.calvin.cs262.pilot.knightrank.TestPOSTPUTSportBackEnd">
-
-    <Button
-        android:id="@+id/send_sport_put"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentEnd="true"
-        android:layout_alignParentBottom="true"
-        android:layout_marginEnd="99dp"
-        android:layout_marginBottom="0dp"
-        android:onClick="put"
-        android:text="PUT" />
-
-    <TextView
-        android:id="@+id/sport_put_id"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginStart="0dp"
-        android:layout_marginTop="95dp"
-        android:text="Sport ID:" />
-
-    <EditText
-        android:id="@+id/input_sport_put_id"
-        android:layout_width="361dp"
-        android:layout_height="52dp"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginStart="0dp"
-        android:layout_marginTop="117dp"
-        android:autofillHints=""
-        android:hint="Enter a id (integer):"
-        android:inputType="text"
-        android:textSize="12sp" />
-
-    <TextView
-        android:id="@+id/post_sport_note_2"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_alignParentEnd="true"
-        android:layout_alignParentBottom="true"
-        android:layout_marginStart="12dp"
-        android:layout_marginTop="336dp"
-        android:layout_marginEnd="8dp"
-        android:layout_marginBottom="196dp"
-        android:text="Bug: Click into any field first time after loading into activity before hitting post, otherwise activity crashes."
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/search_string" />
-
-    <android.support.v7.widget.Toolbar
-        android:id="@+id/my_toolbar"
-        android:layout_width="match_parent"
-        android:layout_height="?attr/actionBarSize"
-        android:background="?attr/colorPrimary"
-        android:elevation="4dp"
-        android:theme="@style/ThemeOverlay.AppCompat.ActionBar"
-        app:popupTheme="@style/ThemeOverlay.AppCompat.Light"/>
-
-    <TextView
-        android:id="@+id/post_put_sport"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentTop="true"
-        android:layout_centerHorizontal="true"
-        android:layout_marginTop="59dp"
-        android:text="POST or PUT a Sport" />
-
-    <TextView
-        android:id="@+id/sport_name"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginStart="0dp"
-        android:layout_marginTop="180dp"
-        android:text="Sport Name:" />
-
-    <TextView
-        android:id="@+id/sport_type"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginStart="0dp"
-        android:layout_marginTop="258dp"
-        android:text="Sport Type:" />
-
-    <EditText
-        android:id="@+id/input_sport_name"
-        android:layout_width="361dp"
-        android:layout_height="52dp"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginStart="0dp"
-        android:layout_marginTop="202dp"
-        android:autofillHints=""
-        android:hint="Enter a name (string):"
-        android:inputType="text"
-        android:textSize="12sp" />
-
-    <EditText
-        android:id="@+id/input_sport_type"
-        android:layout_width="361dp"
-        android:layout_height="52dp"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginStart="1dp"
-        android:layout_marginTop="280dp"
-        android:autofillHints=""
-        android:hint="Enter a type (string)"
-        android:inputType="text"
-        android:textSize="12sp" />
-
-    <Button
-        android:id="@+id/send_sport_post"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentEnd="true"
-        android:layout_alignParentBottom="true"
-        android:onClick="post"
-        android:text="POST" />
-
-    <TextView
-        android:id="@+id/network_connected"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentBottom="true"
-        android:layout_marginStart="16dp"
-        android:layout_marginBottom="18dp"
-        android:text="Network Connectivity" />
-
-    <TextView
-        android:id="@+id/request_status"
-        android:layout_width="344dp"
-        android:layout_height="57dp"
-        android:layout_alignStart="@+id/network_connected"
-        android:layout_alignParentEnd="true"
-        android:layout_alignParentBottom="true"
-        android:layout_marginStart="1dp"
-        android:layout_marginEnd="22dp"
-        android:layout_marginBottom="116dp"
-        android:text="Request Status:" />
-
-</RelativeLayout>
