+++ app/src/main/res/layout/fragment_account_login_instructions.xml
@@ -15,5 +15,4 @@
         android:layout_marginStart="0dp"
         android:layout_marginTop="131dp"
         android:text="Instructions" />
-
 </RelativeLayout>
