--- app/src/main/res/layout/fragment_challenge_confirmation.xml
@@ -1,63 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/fragment_challenge_confirmation_root_layout"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/colorBackground"
-    tools:context=".ChallengeConfirmation">
-
-    <TextView
-        android:id="@+id/challenge_confirmation_textview"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="@string/unresolved_matches"
-        android:textSize="24sp"
-        android:padding="5dp"/>
-
-    <ListView
-        android:id="@+id/challenge_confirmation_listview"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_below="@+id/challenge_confirmation_textview">
-    </ListView>
-
-    <!--
-    <TextView
-        android:id="@+id/score_label"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="10dp"
-        android:layout_marginStart="10dp"
-        android:padding="10dp"
-        android:text="The opponent claims the final score was: "/>
-
-    <EditText
-        android:id="@+id/score1"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="10dp"
-        android:layout_below="@+id/score_label"
-        android:padding="10dp"/>
-
-    <EditText
-        android:id="@+id/score2"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="10dp"
-        android:layout_toEndOf="@+id/score1"
-        android:layout_below="@id/score_label"
-        android:padding="10dp"/>
-
-    <Button
-        android:id="@+id/confirm_button"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentEnd="true"
-        android:layout_alignParentBottom="true"
-        android:text="Confirm"
-        android:padding="10dp"/> -->
-
-</RelativeLayout>
