+++ app/src/main/res/layout/fragment_confirm_match.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/fragment_challenge_confirmation_root_layout"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:background="@color/colorBackground"
+    tools:context=".ConfirmMatch">
+
+    <TextView
+        android:id="@+id/challenge_confirmation_textview"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="@string/unresolved_matches"
+        android:textSize="24sp"
+        android:padding="5dp"/>
+
+    <ListView
+        android:id="@+id/challenge_confirmation_listview"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:layout_below="@+id/challenge_confirmation_textview">
+    </ListView>
+</RelativeLayout>
