+++ app/src/main/res/layout/fragment_declare_match.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorBackground"
-    tools:context=".NewChallenge">
+    tools:context=".DeclareMatch">
 
     <Spinner
         android:id="@+id/activity_spinner"
@@ -57,27 +57,38 @@
         android:padding="10dp"
         android:text="Opponents Score:" />
 
-    <NumberPicker
+    <android.support.design.widget.TextInputLayout
         android:id="@+id/score1"
-        android:layout_width="100dp"
-        android:layout_height="45dp"
+        android:layout_width="150dp"
+        android:layout_height="wrap_content"
         android:layout_below="@+id/score1_label"
         android:layout_alignParentStart="true"
         android:layout_marginStart="10dp"
-        android:padding="10dp">
-    </NumberPicker>
+        android:padding="10dp"
+        android:inputType="numberSigned">
 
-    <NumberPicker
+        <android.support.design.widget.TextInputEditText
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:hint="Enter a score" />
+    </android.support.design.widget.TextInputLayout>
+
+    <android.support.design.widget.TextInputLayout
         android:id="@+id/score2"
-        android:layout_width="100dp"
-        android:layout_height="40dp"
+        android:layout_width="150dp"
+        android:layout_height="wrap_content"
         android:layout_below="@+id/score1_label"
-        android:layout_alignParentEnd="true"
-        android:layout_marginStart="150dp"
+        android:layout_alignParentStart="true"
+        android:layout_marginStart="225dp"
         android:layout_marginEnd="10dp"
-        android:layout_toEndOf="@id/score1"
-        android:padding="10dp">
-    </NumberPicker>
+        android:padding="10dp"
+        android:inputType="numberSigned">
+
+        <android.support.design.widget.TextInputEditText
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:hint="Enter a score" />
+    </android.support.design.widget.TextInputLayout>
 
     <Button
         android:id="@+id/submit"
@@ -87,5 +98,4 @@
         android:layout_alignParentBottom="true"
         android:text="Submit"
         android:padding="10dp"/>
-
 </RelativeLayout>
