+++ app/src/main/res/layout/fragment_leaderboard.xml
@@ -15,7 +15,6 @@
         android:layout_marginTop="5dp"
         android:layout_marginStart="0dp" />
 
-
     <ListView
         android:id="@+id/leaderboard_listview"
         android:layout_width="wrap_content"
@@ -23,5 +22,4 @@
         android:layout_marginTop="40dp"
         android:clickable="false">
     </ListView>
-
 </LinearLayout>
