+++ app/src/main/res/layout/fragment_match_history.xml
@@ -7,8 +7,6 @@
     android:background="@color/colorBackground"
     tools:context=".PastChallenges">
 
-    <!-- TODO: Update blank fragment layout -->
-
     <!--
     <TextView
         android:id="@+id/recent_challenges"
@@ -20,7 +18,6 @@
         android:textColor="@color/colorAuxiliary"
         android:textSize="34sp" />
         -->
-
     <Spinner
         android:id="@+id/past_challenges_spinner"
         android:layout_width="match_parent"
@@ -35,8 +32,5 @@
         android:layout_below="@+id/past_challenges_spinner"
         android:layout_marginLeft="8dp"
         android:layout_marginTop="11dp">
-        <!--
-        android:layout_below="@id/recent_challenges"
-        -->
     </ListView>
 </RelativeLayout>
