--- app/src/main/res/layout/fragment_my_stats.xml
@@ -1,89 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/fragment_my_rankings_root_layout"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/colorBackground"
-    tools:context=".MyStats">
-
-    <!--
-    <TextView
-        android:id="@+id/my_rankings_textview"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="11dp"
-        android:text="@string/title_my_stats"
-        android:textColor="@color/colorAuxiliary"
-        android:textSize="36sp" />
-        -->
-
-    <TextView
-        android:id="@+id/wins_textview"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/wins"
-        android:layout_marginTop="11dp"
-        android:layout_marginStart="5dp"
-        android:textSize="25sp"
-        android:textColor="@color/colorAuxiliary"
-        />
-        <!--
-        android:layout_below="@+id/my_rankings_textview"
-        -->
-
-    <TextView
-        android:id="@+id/win_stat_textview"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="11dp"
-        android:layout_marginStart="2dp"
-        android:layout_toRightOf="@id/wins_textview"
-        android:textSize="25sp"
-        android:text="1"
-        android:textColor="@color/colorAuxiliary"/>
-
-    <TextView
-        android:id="@+id/losses_textview"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/wins_textview"
-        android:layout_marginStart="5dp"
-        android:layout_marginTop="2dp"
-        android:text="@string/losses"
-        android:textSize="25sp"
-        android:textColor="@color/colorAuxiliary"/>
-
-    <TextView
-        android:id="@+id/loss_stat_textview"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/win_stat_textview"
-        android:layout_toRightOf="@id/losses_textview"
-        android:layout_marginStart="2dp"
-        android:layout_marginTop="2dp"
-        android:textSize="25sp"
-        android:text="2"
-        android:textColor="@color/colorAuxiliary"/>
-
-    <TextView
-        android:id="@+id/my_upcoming_matches_textview"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/losses_textview"
-        android:layout_marginStart="5dp"
-        android:layout_marginTop="2dp"
-        android:textSize="25sp"
-        android:text="@string/my_upcoming_matches"/>
-
-    <ListView
-        android:id="@+id/my_stats_upcoming_matches_listview"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/my_upcoming_matches_textview">
-    </ListView>
-
-
-
-</RelativeLayout>
