+++ app/src/main/res/values/strings.xml
@@ -4,8 +4,8 @@
     <string name="title_activity_team_rankings">Team Rankings</string>
     <string name="title_activity_selection">Activity Selection</string>
     <string name="title_my_stats">My Stats</string>
-    <string name="title_challenge_results">Confirm Matches</string>
-    <string name="title_new_challenge">New Challenge</string>
+    <string name="title_confirm_matches">Confirm Matches</string>
+    <string name="title_declare_match">Declare Match</string>
     <string name="splash">splash</string>
     <string name="click_me">Click Me</string>
     <string name="title_screen">Knight Ranker</string>
@@ -64,57 +64,14 @@
         <item>Dalinar Kholin</item>
     </string-array>
 
-    <!-- TODO: Is there a better naming scheme for these -->
-    <string name="button_create_account">Create Account</string>
-    <string name="log_in_button">Log in</string>
-    <string name="title_activity_login">Sign in</string>
-
-    <!-- Strings related to login -->
-    <string name="prompt_email">Email</string>
-    <string name="prompt_password">Password (optional)</string>
-    <string name="action_sign_in">Sign in or register</string>
-    <string name="action_sign_in_short">Sign in</string>
-    <string name="error_invalid_email">This email address is invalid</string>
-    <string name="error_invalid_password">This password is too short</string>
-    <string name="error_incorrect_password">This password is incorrect</string>
-    <string name="error_field_required">This field is required</string>
-    <string name="permission_rationale">"Contacts permissions are needed for providing email
-        completions."
-    </string>
-    <string name="enter_username">Enter Username:</string>
-    <string name="enter_password">Enter Password:</string>
-    <string name="confirm_password">Confirm Password:</string>
-    <string name="return_to_login">Return to Login</string>
-    <string name="create_new_account">Create new account</string>
-    <string name="username">Username:</string>
-    <string name="password">Password:</string>
-    <string name="losses">Losses: </string>
-    <string name="wins">Wins: </string>
-    <string name="location_of_challenge">Location:</string>
-    <string name="activity_of_challenge">Activity:</string>
-
-    <string name="date_of_challenge">Date:</string>
-    <string name="team_challenged">Team:</string>
-    <string name="challenge">Issue Challenge</string>
-    <string name="title_activity_settings">Settings</string>
-
-    <!-- TODO: Remove or change this placeholder text -->
-    <string name="hello_blank_fragment">Hello blank fragment</string>
     <string name="settings">Settings</string>
-
-    <!-- Strings related to Settings -->
-
-    <!-- Example General settings -->
     <string name="pref_header_general">General</string>
-
     <string name="pref_title_social_recommendations">Enable social recommendations</string>
     <string name="pref_description_social_recommendations">Recommendations for people to contact
         based on your message history
     </string>
-
     <string name="pref_title_display_name">Display name</string>
     <string name="pref_default_display_name">John Smith</string>
-
     <string name="pref_title_add_friends_to_messages">Add friends to messages</string>
     <string-array name="pref_example_list_titles">
         <item>Always</item>
@@ -126,10 +83,7 @@
         <item>0</item>
         <item>-1</item>
     </string-array>
-
-    <!-- Example settings for Data & Sync -->
     <string name="pref_header_data_sync">Data &amp; sync</string>
-
     <string name="pref_title_sync_frequency">Sync frequency</string>
     <string-array name="pref_sync_frequency_titles">
         <item>15 minutes</item>
@@ -147,40 +101,19 @@
         <item>360</item>
         <item>-1</item>
     </string-array>
-
-    <string-array name="list_preference_entries">
-        <item>Entry 1</item>
-        <item>Entry 2</item>
-        <item>Entry 3</item>
-    </string-array>
-
-    <string-array name="list_preference_entry_values">
-        <item>1</item>
-        <item>2</item>
-        <item>3</item>
-    </string-array>
-
-    <string-array name="multi_select_list_preference_default_value" />
-
+    <string name="title_activity_settings">Settings</string>
     <string name="pref_title_system_sync_settings">System sync settings</string>
-
-    <!-- Example settings for Notifications -->
     <string name="pref_header_notifications">Notifications</string>
-
     <string name="pref_title_new_message_notifications">New message notifications</string>
-
     <string name="pref_title_ringtone">Ringtone</string>
     <string name="pref_ringtone_silent">Silent</string>
-
     <string name="pref_title_vibrate">Vibrate</string>
     <string name="color_picker">Color picker</string>
     <string name="title_upcoming_challenges">Upcoming Challenges</string>
     <string name="title_view_rankings">Leaderboard</string>
     <string name="my_upcoming_matches">My Upcoming Matches</string>
-    <string name="title_activity_sport_selection">SportSelection</string>
+    <string name="label_activity_selection">SportSelection</string>
     <string name="select_sport_prompt">Select Interests</string>
-    <string name="test_get_backend">Test GET Backend</string>
-
     <!-- Strings for querying via URI-->
     <string name="no_results_found">No Results Found</string>
     <string name="loading_in_process">Loading in process</string>
@@ -192,7 +125,7 @@
     <string name="json_failure">JSON retrieval failure!</string>
     <string name="display_failure">Failed to display results!</string>
     <string name="button_text_fetch">Fetch</string>
-    <string name="sport_selection">Sport Selection</string>
-    <string name="title_past_challenges">Past Matches</string>
+    <string name="activity_selection">Activity Selection</string>
+    <string name="match_history">Match History</string>
     <string name="unresolved_matches">Unresolved Matches</string>
 </resources>
